+++ mayan/apps/events/api_views/event_api_views.py
@@ -1,5 +1,3 @@
-from django.http import Http404
-
 from actstream.models import Action, any_stream
 
 from mayan.apps.rest_api import generics
@@ -7,11 +5,8 @@ from mayan.apps.rest_api.api_view_mixins import (
     ExternalContentTypeObjectAPIViewMixin
 )
 
-from ..classes import EventType, EventTypeNamespace
 from ..permissions import permission_events_view
-from ..serializers import (
-    EventSerializer, EventTypeNamespaceSerializer, EventTypeSerializer
-)
+from ..serializers.event_serializers import EventSerializer
 
 
 class APIObjectEventListView(
@@ -27,73 +22,6 @@ class APIObjectEventListView(
         return any_stream(
             obj=self.get_external_object()
         )
-
-
-class APIEventTypeNamespaceDetailView(generics.RetrieveAPIView):
-    """
-    get: Returns the details of an event type namespace.
-    """
-    serializer_class = EventTypeNamespaceSerializer
-
-    def get_object(self):
-        try:
-            return EventTypeNamespace.get(
-                name=self.kwargs['name']
-            )
-        except KeyError:
-            raise Http404
-
-
-class APIEventTypeNamespaceListView(generics.ListAPIView):
-    """
-    get: Returns a list of all the available event type namespaces.
-    """
-    serializer_class = EventTypeNamespaceSerializer
-    source_queryset = EventTypeNamespace.all()
-
-    def get_serializer_context(self):
-        return {
-            'format': self.format_kwarg,
-            'request': self.request,
-            'view': self
-        }
-
-
-class APIEventTypeNamespaceEventTypeListView(generics.ListAPIView):
-    """
-    get: Returns a list of all the available event types from a namespaces.
-    """
-    serializer_class = EventTypeSerializer
-
-    def get_serializer_context(self):
-        return {
-            'format': self.format_kwarg,
-            'request': self.request,
-            'view': self
-        }
-
-    def get_source_queryset(self):
-        try:
-            return EventTypeNamespace.get(
-                name=self.kwargs['name']
-            ).get_event_types()
-        except KeyError:
-            raise Http404
-
-
-class APIEventTypeListView(generics.ListAPIView):
-    """
-    get: Returns a list of all the available event types.
-    """
-    serializer_class = EventTypeSerializer
-    source_queryset = EventType.all()
-
-    def get_serializer_context(self):
-        return {
-            'format': self.format_kwarg,
-            'request': self.request,
-            'view': self
-        }
 
 
 class APIEventListView(generics.ListAPIView):