+++ mayan/apps/events/api_views/notification_api_views.py
@@ -1,37 +1,21 @@
 from mayan.apps.rest_api import generics
 
-from ..models import Notification
-from ..serializers import NotificationSerializer
+from .api_view_mixins import APIViewMixinNotification
 
 
-class APINotificationDetailView(generics.RetrieveUpdateAPIView):
+class APINotificationDetailView(
+    APIViewMixinNotification, generics.RetrieveUpdateDestroyAPIView
+):
     """
+    delete: Delete the selected notification.
     get: Return the details of the selected notification.
     patch: Edit the selected notification.
     put: Edit the selected notification.
     """
     lookup_url_kwarg = 'notification_id'
-    serializer_class = NotificationSerializer
-
-    def get_source_queryset(self):
-        if self.request.user.is_authenticated:
-            queryset = Notification.objects.filter(user=self.request.user)
-        else:
-            queryset = Notification.objects.none()
-
-        return queryset
 
 
-class APINotificationListView(generics.ListAPIView):
+class APINotificationListView(APIViewMixinNotification, generics.ListAPIView):
     """
     get: Return a list of notifications for the current user.
     """
-    serializer_class = NotificationSerializer
-
-    def get_source_queryset(self):
-        if self.request.user.is_authenticated:
-            queryset = Notification.objects.filter(user=self.request.user)
-        else:
-            queryset = Notification.objects.none()
-
-        return queryset