+++ mayan/apps/events/tests/test_classes.py
@@ -4,7 +4,7 @@ from ..classes import EventModelRegistry, EventType, ModelEventType
 from ..decorators import method_event
 from ..event_managers import EventManagerMethodAfter
 
-from .mixins import EventTypeTestMixin
+from .mixins.event_type_mixins import EventTypeTestMixin
 
 
 class EventManagerTestCase(EventTypeTestMixin, BaseTestCase):