+++ mayan/apps/events/tests/test_notification_api.py
@@ -3,23 +3,58 @@ from rest_framework import status
 from mayan.apps.authentication.events import event_user_logged_out
 from mayan.apps.rest_api.tests.base import BaseAPITestCase
 
-from .mixins import (
-    EventObjectTestMixin, EventTypeTestMixin, NotificationAPIViewTestMixin
-)
+from ..models import Notification
+
+from .mixins.notification_mixins import NotificationAPIViewTestMixin
 
 
 class NotificationAPIViewTestCase(
-    EventObjectTestMixin, EventTypeTestMixin, NotificationAPIViewTestMixin,
-    BaseAPITestCase
+    NotificationAPIViewTestMixin, BaseAPITestCase
 ):
-    def setUp(self):
-        super().setUp()
-        self._create_test_event_type()
-        self._create_test_object_with_event_type_and_permission()
-        self._create_test_event(target=self._test_object)
-        self._create_test_notification()
+    def test_notification_delete_anonymous_api_view(self):
+        self.logout()
 
-    def test_tag_detail_anonymous_api_view(self):
+        notification_count = Notification.objects.count()
+
+        response = self._request_test_notification_delete_api_view()
+        self.assertEqual(response.status_code, status.HTTP_404_NOT_FOUND)
+
+        self.assertEqual(
+            Notification.objects.count(), notification_count
+        )
+
+        events = self._get_test_events()
+        self.assertEqual(events.count(), 2)
+
+        self.assertEqual(events[0].action_object, None)
+        self.assertEqual(events[0].actor, self._test_case_user)
+        self.assertEqual(events[0].target, self._test_object)
+        self.assertEqual(events[0].verb, self._test_event_type.id)
+
+        self.assertEqual(events[1].action_object, None)
+        self.assertEqual(events[1].actor, self._test_case_user)
+        self.assertEqual(events[1].target, self._test_case_user)
+        self.assertEqual(events[1].verb, event_user_logged_out.id)
+
+    def test_notification_delete_api_view(self):
+        notification_count = Notification.objects.count()
+
+        response = self._request_test_notification_delete_api_view()
+        self.assertEqual(response.status_code, status.HTTP_204_NO_CONTENT)
+
+        self.assertEqual(
+            Notification.objects.count(), notification_count - 1
+        )
+
+        events = self._get_test_events()
+        self.assertEqual(events.count(), 1)
+
+        self.assertEqual(events[0].action_object, None)
+        self.assertEqual(events[0].actor, self._test_case_user)
+        self.assertEqual(events[0].target, self._test_object)
+        self.assertEqual(events[0].verb, self._test_event_type.id)
+
+    def test_notification_detail_anonymous_api_view(self):
         self.logout()
 
         response = self._request_test_notification_detail_api_view()
@@ -38,7 +73,7 @@ class NotificationAPIViewTestCase(
         self.assertEqual(events[1].target, self._test_case_user)
         self.assertEqual(events[1].verb, event_user_logged_out.id)
 
-    def test_tag_detail_api_view(self):
+    def test_notification_detail_api_view(self):
         response = self._request_test_notification_detail_api_view()
         self.assertEqual(response.status_code, status.HTTP_200_OK)
 
@@ -55,7 +90,7 @@ class NotificationAPIViewTestCase(
         self.assertEqual(events[0].target, self._test_object)
         self.assertEqual(events[0].verb, self._test_event_type.id)
 
-    def test_tag_edit_anonymous_api_via_patch_view(self):
+    def test_notification_edit_anonymous_api_via_patch_view(self):
         self.logout()
 
         _test_notification_read = self._test_notification.read
@@ -83,7 +118,7 @@ class NotificationAPIViewTestCase(
         self.assertEqual(events[1].target, self._test_case_user)
         self.assertEqual(events[1].verb, event_user_logged_out.id)
 
-    def test_tag_edit_api_via_patch_view(self):
+    def test_notification_edit_api_via_patch_view(self):
         _test_notification_read = self._test_notification.read
 
         response = self._request_test_notification_edit_api_view(verb='patch')
@@ -104,7 +139,7 @@ class NotificationAPIViewTestCase(
         self.assertEqual(events[0].target, self._test_object)
         self.assertEqual(events[0].verb, self._test_event_type.id)
 
-    def test_tag_edit_anonymous_api_via_put_view(self):
+    def test_notification_edit_anonymous_api_via_put_view(self):
         self.logout()
 
         _test_notification_read = self._test_notification.read
@@ -132,7 +167,7 @@ class NotificationAPIViewTestCase(
         self.assertEqual(events[1].target, self._test_case_user)
         self.assertEqual(events[1].verb, event_user_logged_out.id)
 
-    def test_tag_edit_api_via_put_view(self):
+    def test_notification_edit_api_via_put_view(self):
         _test_notification_read = self._test_notification.read
 
         response = self._request_test_notification_edit_api_view(verb='put')