+++ mayan/apps/events/urls.py
@@ -1,9 +1,11 @@
 from django.urls import re_path
 
 from .api_views.event_api_views import (
-    APIEventListView, APIEventTypeListView, APIEventTypeNamespaceDetailView,
+    APIEventListView, APIObjectEventListView
+)
+from .api_views.event_type_api_views import (
+    APIEventTypeListView, APIEventTypeNamespaceDetailView,
     APIEventTypeNamespaceEventTypeListView, APIEventTypeNamespaceListView,
-    APIObjectEventListView
 )
 from .api_views.notification_api_views import (
     APINotificationDetailView, APINotificationListView
@@ -18,7 +20,8 @@ from .views.export_views import (
     EventListExportView, ObjectEventExportView, VerbEventExportView
 )
 from .views.notification_views import (
-    NotificationListView, NotificationMarkRead, NotificationMarkReadAll
+    NotificationDeleteView, NotificationListView, NotificationMarkReadView,
+    NotificationMarkReadAllView
 )
 from .views.subscription_views import (
     EventTypeSubscriptionListView, ObjectEventTypeSubscriptionListView,
@@ -75,13 +78,23 @@ urlpatterns_notification = [
         view=NotificationListView.as_view()
     ),
     re_path(
+        route=r'^user/notifications/multiple/delete/$',
+        name='notification_delete_multiple',
+        view=NotificationDeleteView.as_view()
+    ),
+    re_path(
+        route=r'^user/notifications/(?P<notification_id>\d+)/delete/$',
+        name='notification_delete_single',
+        view=NotificationDeleteView.as_view()
+    ),
+    re_path(
         route=r'^user/notifications/(?P<notification_id>\d+)/mark_read/$',
-        name='notification_mark_read', view=NotificationMarkRead.as_view()
+        name='notification_mark_read', view=NotificationMarkReadView.as_view()
     ),
     re_path(
         route=r'^user/notifications/all/mark_read/$',
         name='notification_mark_read_all',
-        view=NotificationMarkReadAll.as_view()
+        view=NotificationMarkReadAllView.as_view()
     )
 ]
 
@@ -110,7 +123,7 @@ urlpatterns.extend(urlpatterns_events_export)
 urlpatterns.extend(urlpatterns_notification)
 urlpatterns.extend(urlpatterns_subscriptions)
 
-api_urls = [
+api_urls_event_types = [
     re_path(
         route=r'^event_type_namespaces/(?P<name>[-\w]+)/$',
         name='event-type-namespace-detail',
@@ -129,10 +142,20 @@ api_urls = [
     re_path(
         route=r'^event_types/$', name='event-type-list',
         view=APIEventTypeListView.as_view()
-    ),
+    )
+]
+
+api_urls_events = [
     re_path(
         route=r'^events/$', name='event-list', view=APIEventListView.as_view()
     ),
+    re_path(
+        route=r'^objects/(?P<app_label>[-\w]+)/(?P<model_name>[-\w]+)/(?P<object_id>\d+)/events/$',
+        name='object-event-list', view=APIObjectEventListView.as_view()
+    )
+]
+
+api_urls_notifications = [
     re_path(
         route=r'^notifications/$', name='notification-list',
         view=APINotificationListView.as_view()
@@ -140,9 +163,10 @@ api_urls = [
     re_path(
         route=r'^notifications/(?P<notification_id>[0-9]+)/$',
         name='notification-detail', view=APINotificationDetailView.as_view()
-    ),
-    re_path(
-        route=r'^objects/(?P<app_label>[-\w]+)/(?P<model_name>[-\w]+)/(?P<object_id>\d+)/events/$',
-        name='object-event-list', view=APIObjectEventListView.as_view()
     )
 ]
+
+api_urls = []
+api_urls.extend(api_urls_event_types)
+api_urls.extend(api_urls_events)
+api_urls.extend(api_urls_notifications)