+++ HISTORY.rst
@@ -36,6 +36,8 @@
 - Remove vagrant files and make file targets.
 - Update Copyright year from 2011 to 2024.
 - Split URL pattern definitions.
+- Support deleting notifications.
+- Reorganize events app modules.
 
 4.7 (2024-05-07)
 ================